cleaner taxpart

+++ taxalotl/dynamic_partitioning.py
@@ -3,17 +3,21 @@ from __future__ import print_function
 import os
 import re
 
-from peyotl import get_logger, read_as_json, write_as_json
+from peyotl import get_logger, read_as_json, write_as_json, assure_dir_exists
 
 from taxalotl.partitions import (partition_from_mapping,
                                  get_relative_dir_for_partition,
                                  get_all_taxdir_and_misc_uncles,
-                                 get_inverse_misc_non_misc_dir_for_tax)
+                                 get_inverse_misc_non_misc_dir_for_tax,
+                                 get_inp_taxdir,
+                                 get_misc_inp_taxdir,)
+from taxalotl.tax_partition import (TAX_SLICE_CACHE,
+                                    get_taxon_partition,
+                                    TaxonPartition,
+                                    PartitionedTaxDirBase)
 
 _LOG = get_logger(__name__)
-
 _norm_char_pat = re.compile(r'[-a-zA-Z0-9._]')
-
 
 def escape_odd_char(s):
     l = []
@@ -30,45 +34,13 @@ def perform_dynamic_separation(ott_res, part_key, sep_fn):
     top_frag = get_relative_dir_for_partition(part_key)
     general_dynamic_separation(ott_res, top_frag, sep_fn)
 
-
-class TaxonomySliceCache(object):
-    def __init__(self):
-        self._src_filepath_to_vttrs = {}
-
-    def get(self, key):
-        assert isinstance(key, tuple) and len(key) == 2
-        return self._src_filepath_to_vttrs.get(key)
-
-    def __setitem__(self, key, vttrs):
-        assert isinstance(key, tuple) and len(key) == 2
-        assert isinstance(vttrs, VirtualTaxonomyToRootSlice)
-        old_val = self._src_filepath_to_vttrs.get(key)
-        if old_val is not None and old_val is not vttrs:
-            assert False, 'should not be creating a new VTTRS for an inp path!'
-        self._src_filepath_to_vttrs[key] = vttrs
-
-    def flush(self):
-        kv = [(k, v) for k, v in self._src_filepath_to_vttrs.items()]
-        _ex = None
-        for k, v in kv:
-            del self._src_filepath_to_vttrs[k]
-            try:
-                v.flush()
-            except Exception as x:
-                _ex = x
-        if _ex is not None:
-            raise _ex
-
-
 def general_dynamic_separation(ott_res,
                                top_frag,
-                               sep_fn,
-                               tax_slice_cache=None):
+                               sep_fn):
     """
     :param ott_res: resource wrapper for OTT
     :param top_frag: path_frag relative to the top of the partitioned dir
     :param sep_fn:  name of the separations file ('__sep__.json')
-    :param tax_slice_cache: 
     :return: None
     """
     sep_json_filepath = os.path.join(ott_res.partitioned_filepath, top_frag, sep_fn)
@@ -77,69 +49,66 @@ def general_dynamic_separation(ott_res,
         return
     sep_obj = read_as_json(sep_json_filepath)
     cache_creator = False
-    if tax_slice_cache is None:
-        tax_slice_cache = TaxonomySliceCache()
-        cache_creator = True
     try:
         _general_dynamic_separation_from_obj(ott_res,
                                              top_frag,
                                              sep_obj,
-                                             sep_fn,
-                                             tax_slice_cache=tax_slice_cache)
+                                             sep_fn)
     except:
-        if cache_creator:
-            _LOG.exception("clearing TaxonomySliceCache")
-            tax_slice_cache.flush()
+        TAX_SLICE_CACHE.flush()
         raise
 
 
-class VirtualTaxonomyToRootSlice(object):
-    def __init__(self, res, terminal_top_frag, tax_slice_cache, infile_name_list=None):
+class VirtualTaxonomyToRootSlice(PartitionedTaxDirBase):
+    """Represents a taxon for a source, and all of "uncles" back to the root of the taxonomy.
+    """
+    def __init__(self,
+                 res,
+                 fragment):
+        PartitionedTaxDirBase.__init__(self, res, fragment)
         self._taxon_partition = None
+        self._garbage_bin_tax_part = None
         self._to_remove = None
-        self.res = res
-        self.src_id = res.id
-        pd = res.partitioned_filepath
-        if infile_name_list is None:
-            ad = get_all_taxdir_and_misc_uncles(pd, terminal_top_frag, res.id)
-            infile_name_list = [os.path.join(i, res.taxon_filename) for i in ad]
-            infile_name_list = [i for i in infile_name_list if os.path.isfile(i)]
-            if not infile_name_list:
-                raise ValueError("No input taxonomies for {} found in {}".format(res.id, ad))
+        self.misc_uncle = None
+        self._has_flushed = False
+        # find misc_uncle
+        ad = get_all_taxdir_and_misc_uncles(res.partitioned_filepath, fragment, res.id)
+        infile_name_list = [os.path.join(i, res.taxon_filename) for i in ad]
         assert infile_name_list
         term_most = infile_name_list.pop(0)
-        self.own_path = term_most
-        self.inp_dir = os.path.split(term_most)[0]
-        self.cache_key = (self.src_id, self.own_path)
-        self.redundant_to = tax_slice_cache.get(self.cache_key)
-        if self.redundant_to is not None:
-            return
+        assert term_most == self.tax_fp_unpartitioned
         if infile_name_list:
             uncle_fp = infile_name_list[0]
-            self.misc_uncle = tax_slice_cache.get((self.src_id, uncle_fp))
+            self.misc_uncle = TAX_SLICE_CACHE.get((self.src_id, uncle_fp))
             if self.misc_uncle is None:
-                self.misc_uncle = get_virtual_tax_to_root_slice(res,
-                                                                terminal_top_frag=None,
-                                                                tax_slice_cache=tax_slice_cache,
-                                                                infile_name_list=infile_name_list)
-        else:
-            self.misc_uncle = None
-        self._has_flushed = False
-        tax_slice_cache[self.cache_key] = self
-        alt_dir, is_canon = get_inverse_misc_non_misc_dir_for_tax(self.inp_dir, res.id)
-        alt_taxon_filepath = os.path.join(alt_dir, res.taxon_filepath)
-        tax_slice_cache[(self.src_id, alt_taxon_filepath)] = self
-        self.was_unpartitioned = is_canon
-        self.tax_slice_cache = tax_slice_cache
+                uf = os.path.split(self.fragment)[0]
+                mu = get_virtual_tax_to_root_slice(res,
+                                                   fragment=uf)
+                self.misc_uncle = mu
 
-    def separate(self, abs_par_dir, sep_dir_ids_list):
+    def get_vttrs_for_fragment(self, fragment):
+        ck = (VirtualTaxonomyToRootSlice, self.src_id, fragment)
+        return TAX_SLICE_CACHE.get(ck)
+
+    @property
+    def taxon_partition(self):
+        if self._taxon_partition is None:
+            tp = get_taxon_partition(self.res, self.fragment)
+            self._taxon_partition = tp
+        return self._taxon_partition
+
+    def separate(self, fragment_to_partition, sep_dir_ids_list):
         assert self.redundant_to is None
         assert not self._has_flushed
         if not sep_dir_ids_list:
-            _LOG.info("No {} mapping to separate {}".format(self.src_id, abs_par_dir))
+            m = "No {} mapping to separate {}"
+            _LOG.info(m.format(self.src_id, fragment_to_partition))
             return
-        if self._taxon_partition is None:
+        if fragment_to_partition == self.fragment:
+            tp = self.taxon_partition
+            tp.do_partition(sep_dir_ids_list)
             tp, rm_file = partition_from_mapping(self.res,
+                                                 self.fragment,
                                                  sep_dir_ids_list,
                                                  inp_dir=self.inp_dir,
                                                  partition_parsing_fn=self.res.partition_parsing_fn,
@@ -147,23 +116,42 @@ class VirtualTaxonomyToRootSlice(object):
             self._taxon_partition = tp
             assert self._to_remove is None
             self._to_remove = rm_file
+            pd = self.res.partitioned_filepath + '/'
+            for part in tp.partition_el:
+                if part is tp.garbage_bin:
+                    self._garbage_bin_tax_part = part
+                else:
+                    part_dir = part.get_taxon_partition_dir()
+                    assert part_dir.startswith(pd)
+                    part_frag = part_dir[len(pd):]
+                    tf = part.taxon_fp
+                    cached = TAX_SLICE_CACHE.get((self.src_id, tf))
+                    if cached is None:
+                        get_virtual_tax_to_root_slice(self.res,
+                                                      terminal_top_frag=part_frag)
         else:
-            rm_file = self.repartition(sep_dir_ids_list, par_dir=abs_par_dir)
+            rm_file = self.repartition(abs_par_dir, sep_dir_ids_list)
             if rm_file:
                 assert self._to_remove is None or self._to_remove == rm_file
                 self._to_remove = rm_file
         if self.misc_uncle is not None:
-            self.misc_uncle.separate(abs_par_dir, sep_dir_ids_list)
+            self.misc_uncle.separate(fragment_to_partition, sep_dir_ids_list)
 
     def flush(self):
         if self._taxon_partition:
             self._taxon_partition.write()
-        if self._to_remove:
+        if self._to_remove and os.path.exists(self._to_remove):
             _LOG.info("removing pre-partitioned file at {}".format(self._to_remove))
             os.unlink(self._to_remove)
         self._has_flushed = True
 
-    def repartition(self, sep_dir_ids_list, par_dir):
+    def repartition(self, par_dir, sep_dir_ids_list):
+        assert self._garbage_bin_tax_part is not None
+        _LOG.info('repartition: par_dir = {}'.format(par_dir))
+        _LOG.info('repartition: sep_dir_ids_list = {}'.format(sep_dir_ids_list))
+        _LOG.info('repartition: self.own_path = {}'.format(self.own_path))
+        _LOG.info('repartition: garbage_bin.taxon_fp = {}'.format(self._garbage_bin_tax_part.taxon_fp))
+
         raise NotImplementedError("repartition")
 
 """
@@ -185,26 +173,25 @@ def do_new_separation_of_src(res,
 """
 
 
-def get_virtual_tax_to_root_slice(res, terminal_top_frag, tax_slice_cache, infile_name_list=None):
-    vttrs = VirtualTaxonomyToRootSlice(res,
-                                       terminal_top_frag,
-                                       tax_slice_cache,
-                                       infile_name_list=infile_name_list)
-    return vttrs if vttrs.redundant_to is None else vttrs.redundant_to
+def get_virtual_tax_to_root_slice(res,
+                                  fragment):
+    ck = (VirtualTaxonomyToRootSlice, res.id, fragment)
+    c = TAX_SLICE_CACHE.get(ck)
+    if c is not None:
+        return c
+    return VirtualTaxonomyToRootSlice(res, fragment)
 
 
 def _general_dynamic_separation_from_obj(ott_res,
                                          top_frag,
                                          sep_obj,
-                                         sep_fn,
-                                         tax_slice_cache):
+                                         sep_fn):
     """Separtes all sources and handles the recursion. Delegates to do_new_separation_of_src
     
     :param ott_res: 
     :param top_frag: 
     :param sep_obj: 
     :param sep_fn: 
-    :param tax_slice_cache: 
     :return: 
     """
     _LOG.info('breaking {} using {} separators: {}'.format(top_frag,
@@ -223,8 +210,10 @@ def _general_dynamic_separation_from_obj(ott_res,
     taxalotl_config = ott_res.config
     abs_par_dir = os.path.join(ott_res.partitioned_filepath, top_frag)
     for src_id in src_set:
+        if src_id != 'ncbi':
+            continue
         res = taxalotl_config.get_terminalized_res_by_id(src_id)
-        virt_taxon_slice = VirtualTaxonomyToRootSlice(res, top_frag, tax_slice_cache)
+        virt_taxon_slice = get_virtual_tax_to_root_slice(res, top_frag)
         sep_dir_ids_list = []
         for sep_id, i in sep_obj.items():
             t = (sep_id_to_fn[sep_id], i['src_dict'][src_id])
@@ -237,7 +226,8 @@ def _general_dynamic_separation_from_obj(ott_res,
         if not sub:
             continue
         next_frag = os.path.join(top_frag, sep_id_to_fn[sep_id])
-        subdir = os.path.join(abs_par_dir, next_frag)
+        subdir = os.path.join(ott_res.partitioned_filepath, next_frag)
+        assure_dir_exists(subdir)
         subjson = os.path.join(subdir, sep_fn)
         write_as_json(sub, subjson, indent=2, sort_keys=True, separators=(',', ': '))
         _LOG.info('sub separation file written to "{}"'.format(subjson))
@@ -249,5 +239,4 @@ def _general_dynamic_separation_from_obj(ott_res,
         _general_dynamic_separation_from_obj(ott_res=ott_res,
                                              top_frag=next_frag,
                                              sep_obj=next_sep_obj,
-                                             sep_fn=sep_fn,
-                                             tax_slice_cache=tax_slice_cache)
+                                             sep_fn=sep_fn)
